factor out shareable code into own file, make more generic
